Update docs (#593)
* Clean up documentation and tutorials

---------
